dmolinag_df_streaming/requirements.txt
apache-beam[gcp]==2.69.0

# solo librerías externas no incluidas por Beam
protobuf==4.25.3
grpcio==1.62.2
google-auth==2.41.1
google-cloud-logging==3.9.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-dataflow-client==0.8.10
google-cloud-secret-manager==2.20.2

#crypt
pycryptodomex==3.23.0
#--extra-index-url https://us-python.pkg.dev/fif-corp-cl-misc-prod/fif-df-libraries-python/simple/
#cipher-decrypt-encrypt==1.0.0